update master with changes from Sprint 11

+++ src/pages/user-dashboard/user-dashboard.ts
@@ -1,12 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, EventEmitter } from '@angular/core';
 import { IonicPage, NavController, NavParams, PopoverController } from 'ionic-angular';
 import { AuthService } from "../../services/auth.service";
 import { AssessmentService } from "../../services/assessment.service";
 import { TopbarComponent } from "../../components/topbar/topbar";
+import { FileDeleteComponent } from '../../components/file-delete/file-delete';
 import { SettingsPage } from "../settings/settings";
 import { QuestionsPage } from "../questions/questions";
 import { DashboardPage } from "../dashboard/dashboard";
 import { ActionitemsPage } from "../actionitems/actionitems";
+import { EditAssessmentPage } from '../edit-assessment/edit-assessment';
 import { AddTeamMembersPopOverComponent } from "../../components/add-team-members-pop-over/add-team-members-pop-over";
 import { GoogleAnalytics } from '../../application/helpers/GoogleAnalytics';
 
@@ -96,8 +98,6 @@ export class UserDashboardPage {
 		}
   }
 
-
-
 	async getSharedAssessments() {
 		var user;
 		if (this.auth.currentUser()) {
@@ -166,12 +166,37 @@ export class UserDashboardPage {
 
   handleSettings(){ this.navCtrl.push(SettingsPage);}
 
+  async handleEditAssessmentClick(assessmentId) {
+    await this.assessmentService.setCurrentAssessmentId(assessmentId);
+
+    // Go to page to edit assessment.
+    this.navCtrl.push(EditAssessmentPage, {page: 'edit'});
+  }
+
 	toggleMine = () => {this.showMine = !this.showMine;}
 	toggleShared = () => {this.showShared = !this.showShared;}
 
+  /** 
+  *   launch delete popover, pass assessment type
+  *   create an emitter to recieve user response from popover,
+  *   if emitter returns truthy, go use assessment service delete,
+  *   & remove from view
+  *   assessmentId: the assessmentId of the assessment to be deleted
+  */
 	async handleDeleting(assessmentId){
-		var observe =  await this.assessmentService.deleteAssessment(assessmentId);
-		observe.subscribe((data => this.removeAssessmentFromPage(assessmentId)) );
+    var emitter =  new EventEmitter<any>();
+    emitter.subscribe(deleteFile => { 
+      if (deleteFile) {
+        this.assessmentService.deleteAssessment(assessmentId)
+        .then(a => a.toPromise())
+        .catch( err => console.error('cant resolve to Promise'))
+        .then(p => this.removeAssessmentFromPage(assessmentId))
+        .catch( err => console.error('cant remove from page'));
+      }
+    });
+
+    this.popOver.create(FileDeleteComponent, {emitter: emitter, typeToDelete: 'assessment'})
+                .present();
 	}
 
 	presentAddTeamMembersPopOver(assessmentId){
